requirements.txt
autopep8==2.0.2
pydocstyle==6.3.0
pycodestyle==2.10.0
sqlalchemy==2.0.15
sqlalchemy-utils==0.41.1
psycopg2==2.9.6
pyargon2==1.0.7
pyqt6==6.4.2
pyqt6-tools==6.4.2.3.3
matplotlib==3.7.2
testing.postgresql==1.3.0